refactor: Convert to TypeScript
