+++ bin/cmd.js
@@ -9,11 +9,11 @@ const cli = meow(`
 
     Options
 
-      -h, --help                  output usage information
-      -l, --labels <path>         the path to look for the label configuration in. Default: labels.json
-      --token <token>  a GitHub access token (also settable with a GITHUB_ACCESS_TOKEN environment variable)
-      -d, --dry-run               calculate the required label changes but do not apply them
-      -A, --allow-added-labels    allow additional labels in the repo, and don't delete them
+      -h, --help                  [Boolean] output usage information
+      -l, --labels <path>         [Path:String] the path to look for the label configuration in. Default: labels.json
+      --token <token>             [String] a GitHub access token (also settable with a GITHUB_ACCESS_TOKEN environment variable)
+      -d, --dry-run               [Boolean] calculate the required label changes but do not apply them
+      -A, --allow-added-labels    [Boolean] allow additional labels in the repo, and don't delete them
 
 `, {
